Fix: Firefox & Safari Bug: Address Rename doesn’t open modal
On Safari and Firefox, a `click` event is triggered after a `mousedown` & a `mouseup` on the same element. But in the case of the `rename` button, the element start to fadeout before the `mouseup`, so the `click` event is not triggered.
We can then use mousedown as a workaround for Safari & Firefox.