TransactionlistItem,TransactionModal: display BTC swap also if swap tx unknown
This improves how swap transactions are displayed when the BTC transaction
is yet unknown and mainly improves the display of Ledger swaps for
Ledger accounts for which Bitcoin support has not been enabled again
yet after re-import.

Note however, that this only works if the swaps are available in the
swap store. Thus, on a fresh import, the Bitcoins swaps are not shown
as such.