Fix initial BTC sync to not skip address blocks
A computed was watching BTC addresses as well and requesting new ones from Hub when the gap got too small. However, during initial sync this watcher requested addresses in parallel with the history sync, leading to a race condition where the history sync mostly did not receive the next block of addresses, because the watcher already got it.