Mitigate iOS 14.6 IndexedDB issue
Using IndexedDB facilities in Safari always fails in iOS 14.6 on their first
use in a new browser context due to an underlying module lazy loading issue.

Hence, the mitigation waits for the IndexedDB initialisation to complete within
a timeout. If the initialisation isn't completed by then, it is retried until
it succeeds.