Skip recipient details when scanned request contains label or amount
Also adapt recipient-details button text to "Continue" when an amount is already set (either by the scanned request or by the user previously).

This change is mostly triggered by the recent launch of Nimiq on Salamantex PoS terminals, where the UX was confusing as the recipient-details dialog popped up with the button text "Set amount".