Improve WelcomeModal
Store in LocalStorage if modal was already shown, instead of relying on the accounts state after onboarding. This becomes necessary now that the Hub repopulates the Safari/iOS cookie, to prevent the modal always opening when the user only has one account.

Also improves the layout of the network page of the WelcomeModal, whose action button was always placed too high on the screen.

Removed Bitcoin activation modal after onboarding.