Prevent logging into a legacy account not selecting NIM as active currency
This was found by a community member. He was previously logged into a modern account, but his last selected currency was BTC. When he logged into the legacy acocunt, BTC was displayed for that account but not working. There was no way to switch to NIM, as the address list column is used as the legacy-account-upgrade reminder.
Only going to a modern account, switching to NIM and then going to the legacy account fixed it.
We already had a function for this case in the code, but it was not used during after login. Now it is.