swap-kyc-handler: clear sessionStorage before and after requests
As sessionStorage is sand-boxed for separate windows, we use it here for
storing the request and intermediary results to still be available on
reloads and after returning from redirects, without affecting the Wallet
main window and potential other Wallet windows or swaps. However, for the
case that the wallet is configured to use redirects, we still clear the
swap kyc handler storage before and after requests.