swap-kyc-handler: remove dependency on Vue, settings store and composition api
Detect user language from the language cookie instead of the settings store
to get rid of the settings store, Vue  and the composition api as dependencies.

Actually, for using the settings store, we'd have needed to call and await
initStorage, to actually load the store's content from the database, which we
were missing. This would have been introducing just another dependency.

Also, now always let vue i18n return empty strings for missing translations, and
not just in production, to not mistake missing strings displayed as their index
numbers as a bug during dev.