router: refactor BTC activation modal rerouting
Use a navigation guard redirecting to the activation modal for views that
require Bitcoin being activated on the account when it's not activated yet.
In the previous rerouting approach, we loaded the BtcActivationModal during
the lazy view import instead of the original target view. Navigation guards
seem more like the intended way to do this though, and the lazy import
approach had the issue that the imported view was cached, which meant that
we had to reload the page to display the actual target view after Bitcoin
was activated. With this change, we now don't have to reload the pagae
anymore.