Update SignPolygonTransaction request to use Nimiq's OpenGSN contract
This is step 1 of 3. This first step uses Nimiq's OpenGSN contract to execute a token transfer with simultaneous approval. This means a token approval request is embedded into the transaction to Nimiq's OpenGSN contract. 

This transaction does not specify any fees to pay to the OpenGSN RelayHub. That will be added in step 2, as it requires also converting from chainToken to the required amount of USDC swapped via UniSwap.

Step 3 will then be actually sending the transaction through the OpenGSN relay, to not require MATIC anymore from the sender address.