UsdcReceiveModal: improve address tooltip behavior for touch devices
If a copy of the address is triggered via touch, skip the address tooltip that usually opens
on hover and focus, and display the copy tooltip directly. This behavior was already implemented
previously but not very reliable due to race conditions between different events and issues on
devices which have both, touch and a mouse.

Make the behavior more reliable and at the same time simplify the code.