Require relays to have double the required fee as balance during relay selection
This buffer should protect swap users from failing funding or redeem transactions, when the relay relayed another transaction in the meantime which reduced their balance below the fee required by the user's transaction.

Of course, this will reduce the number of eligible relays. We need to monitor if this does not lead to UX problems.