Fix race condition between swap USDC tx-listener and tx-sending
The listener was re-created at every retry, not kept around in the background. This lead to a race condition between the newly created listener and the tx-sending method, usually getting won by the failing of the tx-sending, leading to endless retries although the tx has long been detected in an out-of-scope listener.