router: block activation modals for unsupported account types
Now also block showing the activation modals for legacy accounts and improve
the blocking of the USDC activation modal (and other modals that require USDC
activation, which would previously redirect to the activation modal) for Ledgers.

The USDC activation modal was already previously blocked for Ledgers by closing
itself for Ledger accounts. That approach was pretty hacky though and resulted
in the activation modal having to load and being visible for a moment.