BtcActivationModal: add redirect to Welcome modal
Automatically redirect to the Welcome modal after Btc activation, or offer button to
go there, if already activated, similar to UsdcActivationModal.

This is also especially in preparation to offering optional Bitcoin acivation after
Ledger login again, where the user should be taken to the activation modal first
after login, and then to the Welcome screen afterwards.