hub.ts: automatically offer optional BTC activation on Ledger login
This was already previously supported with 90c4be16 but got broken in 921e8c42.

This also fixes the Welcome modal never being shown for Ledgers since the USDC release
because in hub.ts it was only shown for Keyguard signups or through the UsdcActivationModal
for Keyguard logins, which does not show for Ledger accounts because they don't support
USDC yet. This is now fixed by opening the BtcActivationModal automatically on Ledger
logins, which then leads to the Welcome modal.

Additionally improve some comments related to the welcome modal and activation modals,
and remove the RequestType.ONBOARD redirect response handling because it's unnecessary.