BtcActivationModal,UsdcActivationModal: make skip button open Welcome
Make the skip button redirect to the Welcome Modal, even if the currency is not activated yet,
unless the Welcome modal was already shown, or the main button itself is the 'Check the new intro'
variant that redirects to the Welcome modal.

This is such that the Welcome modal is also shown for Ledger logins or unactivated Keyguard
accounts, even if the user chooses to not go through the activation.