SendModalFooter: support waiting for completely synced Bitcoin history
Completely syncing the Bitcoin history is required for knowing the BTC balance and
available UTXOs.

BtcSendModal was already waiting for syncing of the BTC transaction history in canSign
but did not visually indicate the waiting state with a spinner, which is now the case.

SwapModal and SellCryptoModal were until now not waiting for the complete Bitcoin
history which they now do.

BuyCryptoModal does not need to know the BTC UTXOs and thus does not have to wait
for the complete transaction history.