DataFormatting: change description for Cashlink transactions
"Funding Cashlink" and "Claiming Cashlink" could be confused as still being funded/claimed,
especially by translators, instead of that the transaction WAS funding/claiming the Cashlink.
"Cashlink Funding" and "Cashlink Claiming" makes it slightly more clear that the funding
and claiming is meant, as a substantive. Additionally, a translation hint has been added in
Transifex.