BtcSendModal: improve waiting for address scanned from request link to applied
Don't use a while loop and use a watcher instead, to avoid repeated polling and use the reactivity
system instead to listen for an update. Also don't use a fixed update timeout / iteration count,
because the BtcAddressInput which causes the update does async operations, including fetching
BitcoinJS from the network, whcih can take an undetermined amount of time. Additionally, check
for whether the update is for the expected address, or still from a previous address, to avoid
that our changes get overwritten by the actual update once it arrives.