SendModal: show expiry warning immediately when countdown reaches 0:00
The previous implementation only showed the expiry warning up to .5s after the expiry
because the update interval only runs every 500ms, and up to 1.5s after the countdown
reached 0:00 because the countdown was rounded down (up to 1s for rounding down + .5s
for update interval).
Now, the warning is shown immediately once the countdown reaches 0:00. This can still
be up to .5s after the actual expiry due to the update interval, however that doesn't
really matter as it's not noticeable to the user.