SendModal: show recipient overlay for most payment request links
Now also showing the recipient overlay for payment request links that specify a label or
amount, unless a contact already exists for that recipient address or it's an own or address
book address, or a GoCrypto address. This is for the user to be able to set/edit the label,
if he wants to create a contact, and is also more consistent with the behavior for USDC.
For GoCrypto addresses the contact creation is disabled, because GoCrypto reuses addresses
for different merchants.