PriceChart: show chart based on USD if currency history not supported
Also fixes wrong change percentages, if the history currency is different to the current
currency. This happens for currencies which do not support historic rates, and for which
USD is now used for the chart but could also happen previously, if the history did not
update yet, for example due to coingecko rate limits, after the currency was switched.