Constants: specify enabled fiat currencies via an allow list
Specify the enabled currencies via an allowlist instead of a deny list. This makes the enabled
currencies independent of currency additons to FiatApi, such that the set of enabled currencies
in the Wallet is now the same as before @nimiq/utils v0.10.0, except CRC, GMD and XOF being added.

Also, fixes the prior deny list not being applied in guessUserCurrency and updateExchangeRates
previously. Additionally, the code has been simplified with the allow list.