Show USDC.e notice box only when USDC.e balance is $0.10 or more
When swapping to native USDC, a small remainder will likely stay around, as the fee is over-estimated and not all of it will be used. To not annoy users with showing the box for minimal amounts that cannot be swapped efficiently anyway, we hide the box and USDC.e balance for balances below $0.10.