Use CryptoCompare as provider for historic exchange rates
CoinGecko only allows fetching historic rates of the last 365 day without an api key,
and has strict rate limits which are not suitable for fetching transaction histories
in bulk. CoinGecko is still currently used for continuous updates of the current
exchange rates and price graphs.

Notably, for now, we also reverted using a proxy for CoinGecko, as our API is now
low enough again for using the public api.