BtcTransactions: avoid race conditions in addTransactions
Previously, there was a race condition between addTransactions and calculateFiatAmounts, where
data written to the store by calculateFiatAmounts was overwritten by addTransactions, after it
finished its async code, with data that was retrieved before calculateFiatAmounts ran. Notably,
this could lead to the fiatValue object created by calculateFiatAmounts on transaction outputs
being overwritten/removed, such that attempting to write to it on calculateFiatAmounts finish
results in attempting to write to undefined.
This has been fixed by making addTransactions not async anymore. In NIM and USDC stores, this
was already consciously the case.

Additionally, all async requests in addTransactions are now processed in parallel.