Fiat store: improve CryptoCompare currency limit
CryptoCompare allows 25 currencies to be fetched in a single request, which we want to
limit ourselves to. Previously, we also counted bridged currencies towards this limit,
which resulted in unnecessarily limiting the number of bridged currencies, which don't
need to be restricted, as they don't directly use the CryptoCompare API, and including
less currencies that actually use the CryptoCompare API than allowed as we counted the
bridged currencies towards the 25 currencies limit. This has now been fixed.