Update @nimiq/utils for new CryptoCompare API and use it for historic rates
Note that the old API has been renamed to CryptoCompareLegacy, while the new
API is now uses the name CryptoCompare in FiatApi.

Using CryptoCompare and CryptoCompareLegacy in tandem for load balancing and
increasing our effective rate limits, as both track usages separately. For
historic rates and price charts, use the new CryptoCompare API, as historic
rates in the legacy API seem to be inaccurate as of Nov 2024. For current
rates though, use of the legacy API is fine, as rates between both APIs did
typically not deviate more than 0.1% in our experiments.