Staking events performance: refactor restakingRewards for better performance
Cache accessed values and merge loops, to avoid the unnecessary overhead of Vue's
reactivity system when unnecessarily re-accessing values / events.

This already drastically reduces the computation time of the computed restakingRewards
and cuts the time needed for completing the nextTick by about 50%. The time spent in
Vue's get is cut by 60%-85%. It's still excessively called in the computation of
monthlyRewards in useStakingRewards, which we'll tackle next. On top, the profiling
shows, that it's also excessively used in the computation of transactions in
useTransactionsList, which we might tackle at a later point, as it's a separate issue
from the staking event performance.

New profiling results, based on the same conditions as in the previous commit:

testStakingEventsPerformance: Fetching staking events took 0.238s
testStakingEventsPerformance: Parsing 40748 staking events took 0.332s
testStakingEventsPerformance: setStakingEvents took 1.191s
testStakingEventsPerformance: Vue.nextTick() 1 of 2 took 3.841s
testStakingEventsPerformance: Vue.nextTick() 2 of 2 took 0s
testStakingEventsPerformance: 232 of 40000 staking events are not sorted by date.
testStakingEventsPerformance: sorting (with Date parsing) took 0.399s
testStakingEventsPerformance: tests finished

Most significant `Self time`s in `Bottom-up` view of the profiler:
- get in vue-composition-api.module.js:1014:23: 2,224.7 ms - 29.5 %
- setterHandler in vue-composition-api.module.js:290:36: 1,814.8 ms - 24.1 %
- flushCallbacks in vue.runtime.esm.js:1909:25: 1,158.4 ms - 15.4 %
- (anonymous) in useStakingRewards.ts:54:1 (computed monthlyRewards): 412.3 ms - 5.5 %
- (anonymous) in webpack-internal:///node_modules/idb-keyval/dist/index.js:34:43 (callback for createStore): 341.6 ms - 4.5 %