Staking events performance: avoid double computation of monthlyRewards
A new instance of the computable composable monthlyRewards was created on each invocation
of useStakingRewards, each redoing the heavy computations, resulting in unnecessary double
work between useTransactionList and StakingInfoPage. To avoid this, useStakingRewards
could have been refactored to always return the same computed composable, but instead, I
opted for moving monthlyRewards as a getter into the Staking store, as it fully depends on
nothing but the staking store, and eventually, I might want to couple some computations
more tightly together in the Staking store, for performance improvements or easier caching.

While at it, also remove unused getMonthlyRewards, getCurrentMonthReward and totalRewards
from useStakingRewards. Also move isCurrentMonthAndStakingOngoing and getMonthLabel to
StakingUtils and clean up the code a little bit.

This performance improvement is significant for the StakingInfoPage, but does not show as
an improvement to our previous profiling, as our profiling tests the Wallet main view (the
transaction list).