Staking and syncing performance: don't change entire store objects if only a property changed - experimental!
For example, don't change the entire stakeByAddress object, if only one entry was changed or added.

Changing the entire object results in unnecessarily:
- having to read all old entries to copy them, and thereby invoke their getter reactivity handlers
- having to setup the reactivity handlers again on the new object
- triggering all watchers and computed composables that use the changed object, even if the actually
  changed or added entry does not affect them.

This change is quite experimental and does not seem to be entirely reliable yet in triggering the
the watchers and computed composables, which are meant to be triggered! More testing will be
required to make this production ready. Further refinements to come in the following commits.

In the future, once this is more mature, it will make sense to apply similar techniques to other
stores, where this might be even more relevant, for example the transaction store.

While adapting the Staking store actions in this fashion, also remove unused getters and actions.