Staking and syncing performance: experiments for identifying observed reactivity issues
Implement experiments for identifying the reactivity issues of the previous commit, where changes
to a property via @vue/composition-api's set do not trigger reactive updates, even when they should.
This seems to be happening only for properties not pre-defined in the state, which is however exactly
the corner case according to https://v2.vuejs.org/v2/guide/reactivity.html that set should be handling.
A workaround for correctly receivng reactive updates at usages of the property in watchers and computed
composables has been found, by at each usage making sure the property has been made reactive via set
before accessing it. Additionally, this workaround circumvents the implementation detail of set that it
does not setup reactivity anymore if the property is already defined, even as a non-reactive, regular
property, by deleting the property first, and then redefining them via set.
Next, I'm looking to implement this in a reusable library.