Staking and syncing performance: VueCompositionApiUtils for avoiding reactivity issues
Implement a library VueCompositionApiUtils for avoiding pitfalls of Vue 2's reactivity
system. Most notably, the library provides utilities for setting and accessing a property,
which, on top of what Vue.set and @vue/composition-api's set do, ensure correct reactive
handling, including if the property does not exist yet or exists, but as a regular,
non-reactive property. Apart from that, the library exposes nonReactive for marking
data as non-reactive, and a delete method for deleting a property, as such is not included
in @vue/composition-api. The delete method does unfortunately only notify watchers and
computeds though, which access the value of the property, and not such that only use for
example Object.keys(). Compared to the experimental implementation in removeStake, we now
set deleted entries to undefined, not null, as no actual difference was observed.

Use of this library fixes the reactivity issues observed in the previous commits.