Staking and syncing performance: improve reactivity experiments
Improve the reactivity experiments added two commits ago, and move them to VueCompositionApiUtils.
These new experiments also showed that some of the earlier results were misleading, or misinterpreted.

Analysis: different to what I initially suspected from observations in the actual stores of the Wallet, the set from
@vue/composition-api / Vue.set do in fact work correctly for notifying observers on value changes, and also observers
that use APIs like Object.keys or Object.values when a new property is defined. This works both in plain composables
created with reactive or ref, and also in states of pinia stores. For these cases, set from @vue/composition-api and
Vue.set work just as well as the VueCompositionApiUtils.
What does in fact not work, and what caused the reactivity issues observed in the Wallet, when not reassigning entire
objects when adding a property to enforce a reactive change of the complete object, is: if a property is defined on
the store for the first time by store.patch as opposed to by set or being defined on the initial state, no reactivity
is setup, and also later calls of set on that property will not setup any reactivity anymore. Those issues are fixed
by using our VueCompositionApiUtils by them enforcing reactivity by deleting the existing property, even if it was
already defined by store.patch, however the better way would be to avoid store.patch altogether. Currently, in the
Wallet, all stores are affected, as they all are restored from their persisted state via store.patch, and thus, we
currently always had to do the hack of reassigning entire objects, with the following performance implications:
- running all reactivity getters to copy the old object
- setting up the reactivity handlers again on the new object
- triggering all watchers and computed composables that use the changed object, even if the actually changed or added
  property does not affect them.
Using set to set a property, instead of replacing the entire parent object avoids these performance hits entirely if
writing an existing property, and all but the third one when adding or removing a property. If a new property is
added or removed, set or delete trigger a notification on the parent element, and not just on the property, which
ensures APIs like Object.keys on the object are also re-evaluated correctly, but also lead to watchers and computed
composables being unnecessarily triggered, which use the object, but maybe just a property on that object unrelated
to the added/removed property. In that case, set is thus not so different from assigning a new object, but still has
the other advantages, and clear benefit when setting an already existing property.
To reduce the impact of unnecessarily triggered watchers and computed composables, watchers with a watcher source
that explicitly lists the actual dependencies and triggers the watcher callback only on actual changes can be used,
instead of effect watchers or computed composables. Additionally, repeated watcher invocations could be debounced.