VueCompositionApiUtils: support updating objects in place to avoid unnecessary change notifications
This triggers as little change notifications as possible. For example, if source is equal to target,
no change notifications are triggered, including if source is not the same object as target (no strict
equality), but has the same properties and same values. If only the values of some properties of target
or of a child object changed, only notifications for those properties are triggered. If a property is
added or removed to target or a child object, unfortunately, a change notification on that property's
parent still has to be triggered, as opposed to just on the property.