Staking and syncing performance: avoid unnecessary change notifications in Staking store
Previously, stake or validator changes triggered several unnecessary reactive updates, for example:
setStake() changed the entire stake entry with a new object for an address in stakeByAddress.
The update of stakeByAddress then causes for example a recalculation of the activeStake getter
due to the use of stakeByAddress[activeAddress.value], which was set to a new object in setStake().
The update of activeStake then triggers a recalculation of activeValidator due to the access of
validators.value[activeStake.validator].
The update of activeValidator then triggers a recalculation of restakingRewards due to the access
of activeValidator.address and activeValidator.rewardAddress. This recalculation of restakingRewards is
expensive.

Seeing that these getters actually access only activeStake.validator, activeValidator.address and
activaValidator.rewardAddress, which are all addresses which likely didn't change, the recalculations
are entirely unnecessary, and should only happen if those properties actually changed. However, as
we're replacing the entire stake object in setStake() or the entire apiValidators object in
setApiValidators(), i.e. the parent objects, the getters are being recalculated for their access
of the parent element.

With the use of update and updateObject from our VueCompositionApiUtils, we now make changes in place
to avoid replacing the parent objects, and trigger change notifications only on the properties that
actually changed. If the aforementioned addresses did not change, those getters are now not unnecessarily
triggered anymore.