Staking events performance: fix calculation of historic staking rewards
Rewards of staking events before the selected time range to display in the chart, called
baselineRewards in the now removed code, were included twice, leading to wrong values for
the staking rewards. This was not really visible in the chart, because the graph does not
show its axes and baselineRewards was basically a constant added to each point, such that
the graph looks the same when shifted to fill the canvas. However, avoiding the wrong and
unnecessary calculation of calculateBaselineRewards of course improves performance.