storage.ts: fix reactivity of stores
Remove use of store.patch, because it's broken. It doesn't properly set up reactivity on newly created properties
as it does not use Vue.set to create them. This causes the data to not trigger change notifications, both on
the current change and on future changes, even if Vue.set is used in the future. See experiments and comments
in VueCompositionApiUtils.ts.