VueCompositionApiUtils: add method for patching an object in place
In contrast to update, patch accepts a partial object, and only modifies the properties
of target and its children, that are defined in the passed object. Other properties are
left untouched and are not deleted, which update would do.

Equal to update and updateObject, patch triggers as little change notifications as
possible.