VueCompositionApi: performance optimization for arrays
For regular objects, the setter of Vue's defineReactive will trigger a change notification on the
property if the value did in fact change, based on a === check. For arrays, set triggers a change
notification always on the entire array via array.splice, when setting one of its entries via set.

Because of this behavior for arrays in Vue 2, we can just as well replace the entire array at once,
instead of recursing the array, and save us the effort of detecting the actual changes in the array.
Note that, as already mentioned, for setting properties, set checks whether the value changed at all,
via a === check, and if that is not the case, does not trigger change notifications, so if the array
is a property of an object and its reference remained the same, nothing is triggered, as desired.