Warn users about malicious addresses
Warn the user if he intends to send a transaction to an address detected as malicious,
for example because it has been involved in a hack. The detection runs locally, based
on a block list. Addresses are not censored though, if the user wishes to continue with
the transaction, he can do so, even though it's not encouraged.