Staking events performance: avoid unnecessarily parsing event dates in StakingRewardsModal
Filter events for a month by checking the event's date string against the month string, instead
of parsing the dates, which is much more expensive. Also, previously the dates were unnecessarily
parsed twice for the events in the current month in monthRewardTimestamps. Now, they're only parsed
once and only for the events in the current month. Also, the events are now only iterated once in
monthRewardTimestamps and recomputeFiat, instead of twice.

This makes monthRewardTimestamps ~64% faster and recomputeFiat ~55% (excluding the time needed for
the async network request, which is unaffected).