Staking events performance: remove unused/unreachable StakingRewardsHistoryModal/Page
StakingRewardsHistoryModal had an entry in the router, but was from nowhere linked to.
StakingRewardsHistoryPage was used in StakingRewardsHistoryModal and was declared as
page and as overlay in StakingModal, but there was no code opening this page or overlay.
The overlay was previously accessible via event "next" from StakingInfoPage, but commit
"Add staking rewards to the staking modal" removed it without removing everything else...

As these components are unreachable, it's not worth spending time on optimizing their
performance, and instead we just remove them. Also, there would be plenty to do, because
the performance is so bad that it kills the browser tab.